Added test wallet_dump.py and modified rpc function dumpwallet to return an object with the file name in it.
